Add files via upload
All universities and colleges are mainly focusing on making the process online with a web based application  for the convenience of students and teachers. Our system will be able to create/ join a group for students and coordinators can allocate projects to students and teachers. Thus, a well-developed application for allocation of projects which will help coordinators, teachers and students to keep track of this process throughout the semester. 

1. Projects play an important role in academic growth and development of students. 
2. Getting projects of individual’s choice and interest is also important. 
3. Forming groups, floating forms, going through applications and then allocating projects to groups of students becomes tragic and complicated.
4. Departments like Mechanical, Civil are not that familiar with existing project allocation platforms available.